% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aatemp}
\alias{aatemp}
\title{Annual mean temperatures in Ann Arbor, Michigan}
\format{
\subsection{\code{aatemp}}{

A data frame with 115 rows and 2 columns:
\describe{
\item{year}{Year}
\item{temp}{Temperature in degrees F in Ann Arbor}
}
}
}
\source{
From the package "faraway"
}
\usage{
aatemp
}
\description{
Comes from the U.S. Historical Climatology Network
}
\keyword{datasets}
