% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bank_failures}
\alias{bank_failures}
\title{Bank Data}
\format{
\subsection{\code{aatemp}}{

A data frame with 115 rows and 2 columns:
\describe{
\item{CERT}{cert}
\item{CHCLASS1}{chclass}
...
}
}
}
\source{
From Trevor's files
}
\usage{
bank_failures
}
\description{
Bank failure data
}
\keyword{datasets}
