% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{candydata}
\alias{candydata}
\title{Candy Data}
\format{
\subsection{\code{aatemp}}{

A data frame with ?? rows and ? columns:
\describe{
\item{CERT}{cert}
\item{CHCLASS1}{chclass}
...
}
}
}
\source{
From github
}
\usage{
candydata
}
\description{
Candy Data data
}
\keyword{datasets}
