% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_fuzzybunnies}
\alias{df_fuzzybunnies}
\title{Second data frame in the fuzzybunnies package}
\format{
\subsection{\code{df_fuzzybunnies}}{

A data frame with 5 rows and 2 columns:
\describe{
\item{Name}{list of names}
\item{Age}{list of their ages}
}
}
}
\source{
Brent created in R
}
\usage{
df_fuzzybunnies
}
\description{
Basic data used to test the function df_apply()
}
\keyword{datasets}
