% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diamond_data}
\alias{diamond_data}
\title{Diamond Data}
\format{
\subsection{\code{diamond_data}}{

Data frame with 53940 rows and 7 columns
\describ{
\item{carat}{Unit of Weight}
\item{cut}{Classification of Current Quality}
\item(color){Color Grade}
\item(clarity){Opaqueness}
\item{depth}{Thickness of Diamond}
\item{table}{Largest face of Diamond}
\item{price}{Total Sell Price}
}
}
}
\source{
Github Collection
}
\usage{
diamond_data
}
\description{
Data on diamond statistics and pricing
}
\keyword{datasets}
