% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{marketing_data}
\alias{marketing_data}
\title{Marketing data from kaggle}
\format{
\subsection{\code{marketing_data}}{

A data frame with 2,205 rows and 39 columns:
\describe{
\item{Income}{list of particpant incomes}
\item{Kidhome}{1 if Kid in home, 0 if not}
\item{Teen}{1 if Teen in home, 0 if not}
...
}
}
}
\source{
\url{https://raw.githubusercontent.com/nailson/ifood-data-business-analyst-test/master/ifood_df.csv}
}
\usage{
marketing_data
}
\description{
A public github dataset that contains marketing data about some unknown food company
}
\keyword{datasets}
