% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xyz_fuzzybunnies}
\alias{xyz_fuzzybunnies}
\title{First data frame in the fuzzybunnies package}
\format{
\subsection{\code{xyz_fuzzybunnies}}{

A data frame with 3 rows and 3 columns:
\describe{
\item{x}{values}
\item{y}{more values}
\item{z}{strings}
}
}
}
\source{
Trey created in R
}
\usage{
xyz_fuzzybunnies
}
\description{
Basic data used to test the function df_apply()
}
\keyword{datasets}
