% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_apply.R
\name{df_apply}
\alias{df_apply}
\title{Apply a function to columns of a dataset}
\usage{
df_apply(.data, .fun, .filter, ...)
}
\arguments{
\item{.data}{The dataset}

\item{.fun}{The function to apply}

\item{.filter}{The filter of columns to apply the function}

\item{...}{Additional parameters passed to .fun}
}
\value{
Modified dataset
}
\description{
Apply a function to columns of a dataset
}
\examples{
Name <- c("Jon", "Bill", "Maria", "Ben", "Tina")
Age <- c(23, 41, 32, 58, 26)
df_fuzzybunnies <- data.frame(Name, Age)

increment <- function(x) {
  return(x + 5)
}

df_apply(df_fuzzybunnies, increment, is.numeric)

}
