% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_apply.R
\name{df_apply}
\alias{df_apply}
\title{Apply a function to columns of a dataset}
\usage{
df_apply(.data, .fun, .filter)
}
\arguments{
\item{.data}{The dataset}

\item{.fun}{The function to apply}

\item{.filter}{The filter of columns to apply the function}
}
\value{
Modified dataset
}
\description{
Apply a function to columns of a dataset
}
